Moved all webgl setup out to blur component
