+++ lib/ImageEditor.js
@@ -27,9 +27,6 @@ var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, ge
         step((generator = generator.apply(thisArg, _arguments || [])).next());
     });
 };
-var __importDefault = (this && this.__importDefault) || function (mod) {
-    return (mod && mod.__esModule) ? mod : { "default": mod };
-};
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.ImageEditor = void 0;
 const React = __importStar(require("react"));
@@ -38,16 +35,10 @@ const ControlBar_1 = require("./ControlBar");
 const EditingWindow_1 = require("./EditingWindow");
 const ImageManipulator = __importStar(require("expo-image-manipulator"));
 const Processing_1 = require("./Processing");
-const modal_react_native_web_1 = __importDefault(require("modal-react-native-web"));
 const recoil_1 = require("recoil");
 const Store_1 = require("./Store");
 const OperationBar_1 = require("./OperationBar/OperationBar");
 const noScroll = require("no-scroll");
-const PlatformModal = react_native_1.Platform.OS == "web" ? modal_react_native_web_1.default : react_native_1.Modal;
-// Stop ARIA errors
-if (react_native_1.Platform.OS == "web") {
-    PlatformModal.setAppElement("#root");
-}
 function ImageEditorCore(props) {
     const [imageBounds, setImageBounds] = recoil_1.useRecoilState(Store_1.imageBoundsState);
     const [imageData, setImageData] = recoil_1.useRecoilState(Store_1.imageDataState);
