+++ lib/OperationBar/Blur.js
@@ -18,6 +18,15 @@ var __importStar = (this && this.__importStar) || function (mod) {
     __setModuleDefault(result, mod);
     return result;
 };
+var __awaiter = (this && this.__awaiter) || function (thisArg, _arguments, P, generator) {
+    function adopt(value) { return value instanceof P ? value : new P(function (resolve) { resolve(value); }); }
+    return new (P || (P = Promise))(function (resolve, reject) {
+        function fulfilled(value) { try { step(generator.next(value)); } catch (e) { reject(e); } }
+        function rejected(value) { try { step(generator["throw"](value)); } catch (e) { reject(e); } }
+        function step(result) { result.done ? resolve(result.value) : adopt(result.value).then(fulfilled, rejected); }
+        step((generator = generator.apply(thisArg, _arguments || [])).next());
+    });
+};
 Object.defineProperty(exports, "__esModule", { value: true });
 exports.Blur = void 0;
 const React = __importStar(require("react"));
@@ -26,20 +35,173 @@ const recoil_1 = require("recoil");
 const IconButton_1 = require("../components/IconButton");
 const Store_1 = require("../Store");
 const react_native_slider_1 = require("@miblanchard/react-native-slider");
+const expo_asset_1 = require("expo-asset");
+const vertShader = `
+precision highp float;
+attribute vec2 position;
+varying vec2 uv;
+void main () {
+  uv = position;
+  gl_Position = vec4(1.0 - 2.0 * uv, 0, 1);
+}`;
+const fragShader = `
+precision highp float;
+precision highp int;
+uniform sampler2D texture;
+uniform highp float width;
+uniform highp float height;
+varying vec2 uv;
+uniform highp int radius;
+uniform highp int pass;
+float gauss (float sigma, float x) {
+  float g = (1.0/sqrt(2.0*3.142*sigma*sigma))*exp(-0.5*(x*x)/(sigma*sigma));
+  return g;
+}
+void main () {
+  float f_radius = float(radius);
+  float sigma = ((0.5 * f_radius) + exp(-0.05 * f_radius));
+  // Get the color of the fragment pixel
+  vec4 color = texture2D(texture, vec2(uv.x, uv.y));
+  color *= gauss(sigma, 0.0);
+  for (int i = -50; i <= 50; i++) {
+    if (i >= -radius && i <= radius) {
+      float offset = float(i);
+      // Caclulate the current pixel index
+      float pixelIndex = 0.0;
+      if (pass == 0) {
+        pixelIndex = (uv.y) * height;
+      }
+      else {
+        pixelIndex = uv.x * width;
+      }
+      // Get the neighbouring pixel index
+      pixelIndex += offset;
+      // Normalise the new index back into the 0.0 to 1.0 range
+      if (pass == 0) {
+        pixelIndex /= height;
+      }
+      else {
+        pixelIndex /= width;
+      }
+      // Pad the UV 
+      if (pixelIndex < 0.0) {
+        pixelIndex = 0.0;
+      }
+      if (pixelIndex > 1.0) {
+        pixelIndex = 1.0;
+      }
+      // Get gaussian amplitude
+      float g = gauss(sigma, offset);
+      // Get the color of neighbouring pixel
+      vec4 previousColor = vec4(0.0, 0.0, 0.0, 0.0);
+      if (pass == 0) {
+        previousColor = texture2D(texture, vec2(uv.x, pixelIndex)) * g;
+      }
+      else {
+        previousColor = texture2D(texture, vec2(pixelIndex, uv.y)) * g;
+      }
+      color += previousColor;
+    }
+  }
+  // Return the resulting color
+  gl_FragColor = color;
+}`;
 function Blur() {
     //
     const [, setProcessing] = recoil_1.useRecoilState(Store_1.processingState);
     const [imageData, setImageData] = recoil_1.useRecoilState(Store_1.imageDataState);
     const [, setEditingMode] = recoil_1.useRecoilState(Store_1.editingModeState);
     const [glContext, setGLContext] = recoil_1.useRecoilState(Store_1.glContextState);
-    const [glProgram, setGLProgram] = recoil_1.useRecoilState(Store_1.glProgramState);
     const [blur, setBlur] = React.useState(1);
+    const [glProgram, setGLProgram] = React.useState(null);
     const onClose = () => {
         // If closing reset the image back to its original
-        setGLProgram(null);
         setGLContext(null);
         setEditingMode("operation-select");
     };
+    React.useEffect(() => {
+        if (glContext !== null) {
+            const setupGL = () => __awaiter(this, void 0, void 0, function* () {
+                // Load in the asset and get its height and width
+                const gl = glContext;
+                let asset = expo_asset_1.Asset.fromURI(imageData.uri);
+                yield asset.downloadAsync();
+                asset.height = imageData.height;
+                asset.width = imageData.width;
+                if (asset.width && asset.height) {
+                    // Setup the shaders for our GL context so it draws from texImage2D
+                    const vert = gl.createShader(gl.VERTEX_SHADER);
+                    const frag = gl.createShader(gl.FRAGMENT_SHADER);
+                    if (vert && frag) {
+                        // Set the source of the shaders and compile them
+                        gl.shaderSource(vert, vertShader);
+                        gl.compileShader(vert);
+                        gl.shaderSource(frag, fragShader);
+                        gl.compileShader(frag);
+                        // Create a WebGL program so we can link the shaders together
+                        const program = gl.createProgram();
+                        if (program) {
+                            // Attach both the vertex and frag shader to the program
+                            gl.attachShader(program, vert);
+                            gl.attachShader(program, frag);
+                            // Link the program - ensures that vetex and frag shaders are compatible
+                            // with each other
+                            gl.linkProgram(program);
+                            // Tell GL we ant to now use this program
+                            gl.useProgram(program);
+                            // Create a buffer on the GPU and assign its type as array buffer
+                            const buffer = gl.createBuffer();
+                            gl.bindBuffer(gl.ARRAY_BUFFER, buffer);
+                            // Create the verticies for WebGL to form triangles on the screen
+                            // using the vertex shader which forms a square or rectangle in this case
+                            const verts = new Float32Array([
+                                -1,
+                                -1,
+                                1,
+                                -1,
+                                1,
+                                1,
+                                -1,
+                                -1,
+                                -1,
+                                1,
+                                1,
+                                1,
+                            ]);
+                            // Actually pass the verticies into the buffer and tell WebGL this is static
+                            // for optimisations
+                            gl.bufferData(gl.ARRAY_BUFFER, verts, gl.STATIC_DRAW);
+                            // Get the index in memory for the position attribute defined in the
+                            // vertex shader
+                            const positionAttrib = gl.getAttribLocation(program, "position");
+                            gl.enableVertexAttribArray(positionAttrib); // Enable it i guess
+                            // Tell the vertex shader how to process this attribute buffer
+                            gl.vertexAttribPointer(positionAttrib, 2, gl.FLOAT, false, 0, 0);
+                            // Fetch an expo asset which can passed in as the source for the
+                            // texImage2D
+                            // Create some space in memory for a texture
+                            const texture = gl.createTexture();
+                            // Set the active texture to the texture 0 binding (0-30)
+                            gl.activeTexture(gl.TEXTURE0);
+                            // Bind the texture to WebGL stating what type of texture it is
+                            gl.bindTexture(gl.TEXTURE_2D, texture);
+                            // Set some parameters for the texture
+                            gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MAG_FILTER, gl.LINEAR);
+                            gl.texParameteri(gl.TEXTURE_2D, gl.TEXTURE_MIN_FILTER, gl.LINEAR);
+                            // Then set the data of this texture using texImage2D
+                            gl.texImage2D(gl.TEXTURE_2D, 0, gl.RGBA, gl.RGBA, gl.UNSIGNED_BYTE, asset);
+                            // Set a bunch of uniforms we want to pass into our fragement shader
+                            gl.uniform1i(gl.getUniformLocation(program, "texture"), 0);
+                            gl.uniform1f(gl.getUniformLocation(program, "width"), asset.width);
+                            gl.uniform1f(gl.getUniformLocation(program, "height"), asset.height);
+                            setGLProgram(program);
+                        }
+                    }
+                }
+            });
+            setupGL();
+        }
+    }, [glContext]);
     React.useEffect(() => {
         const gl = glContext;
         const program = glProgram;
