+++ lib/OperationBar/OperationSelection.js
@@ -94,7 +94,7 @@ const styles = react_native_1.StyleSheet.create({
         justifyContent: "space-around",
     },
     modeButton: {
-        height: 70,
+        height: 80,
         flex: 1,
         justifyContent: "center",
         alignItems: "center",
