+++ package.json
@@ -24,6 +24,10 @@
     "@types/react-dom": "^16.9.8",
     "@types/react-native": "^0.62.11",
     "babel-preset-expo": "~8.1.0",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-native": "https://github.com/expo/react-native/archive/sdk-39.0.0.tar.gz",
+    "react-native-web": "~0.13.7",
     "typescript": "^3.9.3"
   },
   "dependencies": {
@@ -36,11 +40,6 @@
     "expo-image-manipulator": "~8.3.0",
     "modal-enhanced-react-native-web": "^0.2.0",
     "modal-react-native-web": "^0.2.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-native": "https://github.com/expo/react-native/archive/sdk-39.0.0.tar.gz",
-    "react-native-screens": "~2.10.1",
-    "react-native-web": "~0.13.7",
     "recoil": "^0.0.10"
   }
 }
